+++ mypyc/irbuild/specialize.py
@@ -22,16 +22,15 @@ from mypyc.ir.ops import (
 )
 from mypyc.ir.rtypes import (
     RType, RTuple, str_rprimitive, list_rprimitive, dict_rprimitive, set_rprimitive,
-    bool_rprimitive, is_dict_rprimitive, is_list_rprimitive, is_tuple_rprimitive
+    bool_rprimitive, is_dict_rprimitive
 )
-from mypyc.primitives.registry import CFunctionDescription
 from mypyc.primitives.dict_ops import dict_keys_op, dict_values_op, dict_items_op
 from mypyc.primitives.tuple_ops import new_tuple_set_item_op
 from mypyc.primitives.list_ops import list_set_item_op
 from mypyc.irbuild.builder import IRBuilder
 from mypyc.irbuild.for_helpers import (
     translate_list_comprehension, translate_set_comprehension,
-    comprehension_helper, for_loop_helper_with_index,
+    comprehension_helper, preallocate_space_helper
 )
 
 
@@ -160,7 +159,8 @@ def translate_safe_generator_call(
         else:
             if len(expr.args) == 1 and callee.fullname == "builtins.tuple":
                 val = preallocate_space_helper(builder, expr.args[0],
-                                               empty_op_llbuilder=builder.builder.new_tuple_with_length,
+                                               empty_op_llbuilder=
+                                                   builder.builder.new_tuple_with_length,
                                                set_item_op=new_tuple_set_item_op)
                 if val is not None:
                     return val
@@ -169,30 +169,6 @@ def translate_safe_generator_call(
                 expr, callee,
                 ([translate_list_comprehension(builder, expr.args[0])]
                     + [builder.accept(arg) for arg in expr.args[1:]]))
-    return None
-
-
-def preallocate_space_helper(builder: IRBuilder,
-                             gen: GeneratorExpr,
-                             empty_op_llbuilder: Callable[[Value, int], Value],
-                             set_item_op: CFunctionDescription) -> Optional[Value]:
-    """Currently we only optimize for simplest generator expression"""
-    if len(gen.sequences) == 1 and len(gen.condlists[0]) == 0:
-        rtype = builder.node_type(gen.sequences[0])
-        if is_list_rprimitive(rtype) or is_tuple_rprimitive(rtype):
-
-            length = builder.builder.builtin_len(builder.accept(gen.sequences[0]),
-                                                 gen.line, use_pyssize_t=True)
-            target_op = empty_op_llbuilder(length, gen.line)
-
-            def set_item(item_index: Value) -> None:
-                e = builder.accept(gen.left_expr)
-                builder.call_c(set_item_op, [target_op, item_index, e], gen.line)
-
-            for_loop_helper_with_index(builder, gen.indices[0], gen.sequences[0],
-                                       set_item, gen.line)
-
-            return target_op
     return None
